wasm-runner/controller/requirements.txt
kubernetes
shortuuid
kafka-python == 2.0.2
dnspython